requirements.txt
dash
dash-bootstrap-components
plotly
wfdb
numpy
scipy
tensorflow
pyts
Pillow
